Создание абстрактного класса. Настройка наследования компонентов от абстрактного класса
Пока не реализованы компоненты показа пустого списка маршрута и показа окна при добавлении нового пункта маршрута.
