+++ src/view/trip-info.js
@@ -1,4 +1,4 @@
-import { createElement } from '../utils/render.js';
+import AbstractView from './abstract-view.js';
 
 const createTripInfoTemplate = () => (
   `<section class="trip-main__trip-info  trip-info">
@@ -14,22 +14,8 @@ const createTripInfoTemplate = () => (
   </section>`
 );
 
-export default class TripInfoView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class TripInfoView extends AbstractView {
   get template() {
     return createTripInfoTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
