+++ src/view/trip-navigation.js
@@ -1,4 +1,4 @@
-import { createElement } from '../utils/render.js';
+import AbstractView from './abstract-view.js';
 
 const createTripNavigationTemplate = () => (
   `<nav class="trip-controls__trip-tabs  trip-tabs">
@@ -7,22 +7,8 @@ const createTripNavigationTemplate = () => (
   </nav>`
 );
 
-export default class TripNavigationView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class TripNavigationView extends AbstractView {
   get template() {
     return createTripNavigationTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
