+++ src/view/trip-sort.js
@@ -1,4 +1,4 @@
-import { createElement } from '../utils/render.js';
+import AbstractView from './abstract-view.js';
 
 const createTripSortTemplate = () => (
   `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
@@ -29,22 +29,8 @@ const createTripSortTemplate = () => (
   </form>`
 );
 
-export default class TripSortView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class TripSortView extends AbstractView {
   get template() {
     return createTripSortTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
